# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathashaala Dining Seat Arrangement API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type PersonRole = (typeof PersonRole)[keyof typeof PersonRole];

export const PersonRole = {
  student: "student",
  teacher: "teacher",
  non_teaching_staff: "non_teaching_staff",
} as const;

export type StudentGroup = (typeof StudentGroup)[keyof typeof StudentGroup];

export const StudentGroup = {
  junior_mag: "junior_mag",
  senior_mag: "senior_mag",
  grade_9: "grade_9",
  grade_10: "grade_10",
  grade_11: "grade_11",
  grade_12: "grade_12",
} as const;

export interface Person {
  id: number;
  name: string;
  role: PersonRole;
  studentGroup?: StudentGroup | null;
}

export interface PeopleListResponse {
  people: Person[];
}

export interface CreatePersonRequest {
  name: string;
  role: PersonRole;
  studentGroup?: StudentGroup | null;
}

export interface BulkCreatePeopleRequest {
  people: CreatePersonRequest[];
}

export interface BulkCreatePeopleResponse {
  created: number;
  skipped: number;
  errors: string[];
}

export interface Config {
  /** @minimum 1 */
  tableCount: number;
  /** @minimum 1 */
  maxTeachersPerTable: number;
  /** @minimum 2 */
  seatsPerTable: number;
}

export interface UpdateConfigRequest {
  /** @minimum 1 */
  tableCount?: number;
  /** @minimum 1 */
  maxTeachersPerTable?: number;
  /** @minimum 2 */
  seatsPerTable?: number;
}

export interface GroupMember {
  personId: number;
  personName: string;
}

export interface Group {
  id: number;
  name: string;
  /**
   * 0 = don't care, 100 = should never be together
   * @minimum 0
   * @maximum 100
   */
  friendshipLevel: number;
  members: GroupMember[];
}

export interface GroupsListResponse {
  groups: Group[];
}

export interface CreateGroupRequest {
  name: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  friendshipLevel: number;
  /** @minItems 2 */
  memberPersonIds: number[];
}

export interface UpdateGroupRequest {
  name?: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  friendshipLevel?: number;
  /** @minItems 2 */
  memberPersonIds?: number[];
}

export interface TableArrangement {
  tableNumber: number;
  seats: Person[];
}

export interface WeekArrangement {
  weekNumber: number;
  tables: TableArrangement[];
}

export interface GenerateArrangementsResponse {
  weeks: WeekArrangement[];
  message: string;
}