# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathashaala Dining Seat Arrangement API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all people
 */
export const ListPeopleResponse = zod.object({
  people: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      role: zod.enum(["student", "teacher", "non_teaching_staff"]),
      studentGroup: zod
        .enum([
          "junior_mag",
          "senior_mag",
          "grade_9",
          "grade_10",
          "grade_11",
          "grade_12",
        ])
        .nullish(),
    }),
  ),
});

/**
 * @summary Add a person
 */
export const CreatePersonBody = zod.object({
  name: zod.string(),
  role: zod.enum(["student", "teacher", "non_teaching_staff"]),
  studentGroup: zod
    .enum([
      "junior_mag",
      "senior_mag",
      "grade_9",
      "grade_10",
      "grade_11",
      "grade_12",
    ])
    .nullish(),
});

/**
 * @summary Bulk add people from parsed Excel data
 */
export const BulkCreatePeopleBody = zod.object({
  people: zod.array(
    zod.object({
      name: zod.string(),
      role: zod.enum(["student", "teacher", "non_teaching_staff"]),
      studentGroup: zod
        .enum([
          "junior_mag",
          "senior_mag",
          "grade_9",
          "grade_10",
          "grade_11",
          "grade_12",
        ])
        .nullish(),
    }),
  ),
});

export const BulkCreatePeopleResponse = zod.object({
  created: zod.number(),
  skipped: zod.number(),
  errors: zod.array(zod.string()),
});

/**
 * @summary Remove a person
 */
export const DeletePersonParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get table configuration
 */

export const getConfigResponseSeatsPerTableMin = 2;

export const GetConfigResponse = zod.object({
  tableCount: zod.number().min(1),
  maxTeachersPerTable: zod.number().min(1),
  seatsPerTable: zod.number().min(getConfigResponseSeatsPerTableMin),
});

/**
 * @summary Update table configuration
 */

export const updateConfigBodySeatsPerTableMin = 2;

export const UpdateConfigBody = zod.object({
  tableCount: zod.number().min(1).optional(),
  maxTeachersPerTable: zod.number().min(1).optional(),
  seatsPerTable: zod.number().min(updateConfigBodySeatsPerTableMin).optional(),
});

export const updateConfigResponseSeatsPerTableMin = 2;

export const UpdateConfigResponse = zod.object({
  tableCount: zod.number().min(1),
  maxTeachersPerTable: zod.number().min(1),
  seatsPerTable: zod.number().min(updateConfigResponseSeatsPerTableMin),
});

/**
 * @summary List all friendship groups
 */
export const listGroupsResponseGroupsItemFriendshipLevelMin = 0;
export const listGroupsResponseGroupsItemFriendshipLevelMax = 100;

export const ListGroupsResponse = zod.object({
  groups: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      friendshipLevel: zod
        .number()
        .min(listGroupsResponseGroupsItemFriendshipLevelMin)
        .max(listGroupsResponseGroupsItemFriendshipLevelMax)
        .describe("0 = don't care, 100 = should never be together"),
      members: zod.array(
        zod.object({
          personId: zod.number(),
          personName: zod.string(),
        }),
      ),
    }),
  ),
});

/**
 * @summary Create a friendship group
 */
export const createGroupBodyFriendshipLevelMin = 0;
export const createGroupBodyFriendshipLevelMax = 100;

export const createGroupBodyMemberPersonIdsMin = 2;

export const CreateGroupBody = zod.object({
  name: zod.string(),
  friendshipLevel: zod
    .number()
    .min(createGroupBodyFriendshipLevelMin)
    .max(createGroupBodyFriendshipLevelMax),
  memberPersonIds: zod
    .array(zod.number())
    .min(createGroupBodyMemberPersonIdsMin),
});

/**
 * @summary Update a friendship group
 */
export const UpdateGroupParams = zod.object({
  id: zod.coerce.number(),
});

export const updateGroupBodyFriendshipLevelMin = 0;
export const updateGroupBodyFriendshipLevelMax = 100;

export const updateGroupBodyMemberPersonIdsMin = 2;

export const UpdateGroupBody = zod.object({
  name: zod.string().optional(),
  friendshipLevel: zod
    .number()
    .min(updateGroupBodyFriendshipLevelMin)
    .max(updateGroupBodyFriendshipLevelMax)
    .optional(),
  memberPersonIds: zod
    .array(zod.number())
    .min(updateGroupBodyMemberPersonIdsMin)
    .optional(),
});

export const updateGroupResponseFriendshipLevelMin = 0;
export const updateGroupResponseFriendshipLevelMax = 100;

export const UpdateGroupResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  friendshipLevel: zod
    .number()
    .min(updateGroupResponseFriendshipLevelMin)
    .max(updateGroupResponseFriendshipLevelMax)
    .describe("0 = don't care, 100 = should never be together"),
  members: zod.array(
    zod.object({
      personId: zod.number(),
      personName: zod.string(),
    }),
  ),
});

/**
 * @summary Delete a friendship group
 */
export const DeleteGroupParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Generate 10 weekly seating arrangements and return Excel file URLs
 */
export const GenerateArrangementsResponse = zod.object({
  weeks: zod.array(
    zod.object({
      weekNumber: zod.number(),
      tables: zod.array(
        zod.object({
          tableNumber: zod.number(),
          seats: zod.array(
            zod.object({
              id: zod.number(),
              name: zod.string(),
              role: zod.enum(["student", "teacher", "non_teaching_staff"]),
              studentGroup: zod
                .enum([
                  "junior_mag",
                  "senior_mag",
                  "grade_9",
                  "grade_10",
                  "grade_11",
                  "grade_12",
                ])
                .nullish(),
            }),
          ),
        }),
      ),
    }),
  ),
  message: zod.string(),
});

/**
 * @summary Download a specific week's Excel arrangement
 */
export const downloadArrangementPathWeekIndexMin = 0;
export const downloadArrangementPathWeekIndexMax = 9;

export const DownloadArrangementParams = zod.object({
  weekIndex: zod.coerce
    .number()
    .min(downloadArrangementPathWeekIndexMin)
    .max(downloadArrangementPathWeekIndexMax),
});