# lib/api-zod/src/generated/types/bulkCreatePeopleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathashaala Dining Seat Arrangement API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePersonRequest } from "./createPersonRequest";

export interface BulkCreatePeopleRequest {
  people: CreatePersonRequest[];
}