# lib/api-zod/src/generated/types/bulkCreatePeopleResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathashaala Dining Seat Arrangement API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkCreatePeopleResponse {
  created: number;
  skipped: number;
  errors: string[];
}