# lib/api-zod/src/generated/types/createPersonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathashaala Dining Seat Arrangement API
 * OpenAPI spec version: 0.1.0
 */
import type { PersonRole } from "./personRole";
import type { StudentGroup } from "./studentGroup";

export interface CreatePersonRequest {
  name: string;
  role: PersonRole;
  studentGroup?: StudentGroup | null;
}