# lib/api-zod/src/generated/types/generateArrangementsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathashaala Dining Seat Arrangement API
 * OpenAPI spec version: 0.1.0
 */
import type { WeekArrangement } from "./weekArrangement";

export interface GenerateArrangementsResponse {
  weeks: WeekArrangement[];
  message: string;
}