# lib/api-zod/src/generated/types/group.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathashaala Dining Seat Arrangement API
 * OpenAPI spec version: 0.1.0
 */
import type { GroupMember } from "./groupMember";

export interface Group {
  id: number;
  name: string;
  /**
   * 0 = don't care, 100 = should never be together
   * @minimum 0
   * @maximum 100
   */
  friendshipLevel: number;
  members: GroupMember[];
}